backend/requirements_chatbot.txt
# Robo-Advisor Chatbot Dependencies

# Core FastAPI and async support
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Pydantic and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAI integration
openai==1.3.7

# Vector database
qdrant-client==1.7.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 